split feed and item tests into own files
